Re-Added GDT Creator to mlMainWindow
Didn't notice it would stop compilation. I might swap it around later.
