+++ ModLauncher.vcxproj
@@ -306,7 +306,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="dvar.cpp" />
-    <ClCompile Include="GDTCreator.cpp" />
     <ClCompile Include="GeneratedFiles\Debug\moc_mlMainWindow.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='External_Release|Win32'">true</ExcludedFromBuild>
@@ -476,7 +475,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="dvar.h" />
-    <ClInclude Include="GDTCreator.h" />
     <CustomBuild Include="syntax.h">
       <AdditionalInputs Condition="'$(Configuration)|$(Platform)'=='External_Release|Win32'">$(QTDIR)\bin\moc.exe;%(FullPath)</AdditionalInputs>
       <Message Condition="'$(Configuration)|$(Platform)'=='External_Release|Win32'">Moc%27ing syntax.h...</Message>
