+++ ModLauncher.vcxproj.filters
@@ -67,9 +67,6 @@
     <ClCompile Include="GeneratedFiles\Release\moc_syntax.cpp">
       <Filter>Generated Files\Release</Filter>
     </ClCompile>
-    <ClCompile Include="GDTCreator.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="ModLauncher.rc">
@@ -95,9 +92,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="dvar.h">
-      <Filter>Source Files</Filter>
-    </ClInclude>
-    <ClInclude Include="GDTCreator.h">
       <Filter>Source Files</Filter>
     </ClInclude>
   </ItemGroup>
