+++ mlMainWindow.h
@@ -229,3 +229,16 @@ public:
 	Export2BinGroupBox(QWidget *parent, mlMainWindow* parent_window);
 };
 
+class GDTCreatorGroupBox : public QGroupBox
+{
+private:
+	mlMainWindow* parentWindow;
+
+protected:
+	void dragEnterEvent(QDragEnterEvent* event);
+	void dragLeaveEvent(QDragLeaveEvent* event);
+	void dropEvent(QDropEvent *event);
+
+public:
+	GDTCreatorGroupBox(QWidget *parent, mlMainWindow* parent_window);
+};
