+++ stdafx.h
@@ -22,9 +22,7 @@
 #include "steam_api.h"
 #include "dvar.h"
 #include "syntax.h"
-#include "GDTCreator.h"
 
 class mlMainWindow;
 class mlExport2BinWidget;
 class Syntax;
-class GDTCreatorGroupBox;
